+++ lukefi/metsi/domain/sim_ops.py
@@ -9,7 +9,7 @@ from lukefi.metsi.domain.data_collection.marshalling import (
     collect_felled_tree_properties,
     report_period
 )
-from lukefi.metsi.domain.natural_processes.grow_acta import grow_acta
+from lukefi.metsi.domain.natural_processes.grow_acta import grow_acta, grow_acta_vectorized
 from lukefi.metsi.domain.natural_processes.grow_metsi import grow_metsi
 from lukefi.metsi.domain.natural_processes.grow_motti_dll import grow_motti_dll
 from lukefi.metsi.domain.forestry_operations.thinning import (
@@ -25,6 +25,7 @@ from lukefi.metsi.sim.operations import do_nothing
 
 __all__ = ['grow_metsi',
            'grow_acta',
+           'grow_acta_vectorized',
            'grow_motti_dll',
            'thinning_from_below',
            'thinning_from_above',