+++ CHANGELOG.md
@@ -4,6 +4,37 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [2.3.6] - 2025-08-28
+
+### Changed
+
+- Support for python version >=3.12
+
+## [2.3.5] - 2025-08-28
+
+### Changed
+
+- Old data is now deleted after vectorization
+
+## [2.3.4] - 2025-08-27
+
+### Fixed
+
+- Minor code quality improvements
+- Pylint configuration updated for changed MetsiGrow path
+
+## [2.3.3] - 2025-08-27
+
+### Added
+
+- Added grow_acta_vectorized - a version of grow_acta that operates on SoA data
+
+## [2.3.2] - 2025-08-26
+
+### Fixed
+
+- Updated examples to work with earlier control changes
+
 ## [2.3.1] - 2025-08-26
 
 ### Added